Initial import into git
Revision IX (revision9, r9, etc.) is a server engine which manages 3D
space.  Each server has a set of skills, and each actor within the server
has a level in some of those skills.  The skills, and also the physics
engine of each server is intended to be modular and changeable.