Sockets now use the new-style listen_socket interface
Since the new listen_socket class hierarchy now has threading and a lot
of nice features, we're now using them to create the listening sockets
in the main program.

Also updated the makefile for the classes subdirectory, to get it to at
least build successfully.

Removed the sockets item from the TODO list, and added a few new items.