Small tweaks to get everything to build
Most of the server makefiles needed tweaking to get things to build
correctly.  Converted some of the shared libs (which seem to link
strangely or not at all on OSX) into archive files, which link with
no problems.

I had neglected to include the files for the syslog streambuf, both
of which needed some tweaking to compile correctly.  Inserted a couple
of console stubs into server.cc, cleaned out some nonexistent prototypes
from server.h.  Moved the config default value initializers from the
header to the implementation file.  Added the binary to .gitignore.