Updated octree to be a class
Previously, the octree was a struct with a collection of routines which
built the various parts of the data.  Now it's an honest-to-goodness class,
which has a proper constructor, and handles subtrees as if they were full
trees themselves.

Next step is to decide how to include them within the zone.  Each of the
steps for each dimension needs its own octree, which implies a multi-
dimensional array, but the C++ array type is allocated at compile time,
rather than runtime.  A triple-nested vector may be the way to go.