Added the new octrees into the zone
We've now got a three-level vector of Octree pointers in the zone object.
We don't do anything with them yet, but they're there.  We may consider
making some reserve calls on the various levels of vector within either
the constructor or the start method, to eliminate possible memory
craziness while the game is up and running.

The polygon class was no longer used, and will likely never be used again,
so it's gone.