Moved the access pool into the listen_socket
Having a single global access pool (logins and logouts) was far from ideal,
so it is now inside the listen_socket, where the connections come in the
first place.  The access.cc file is gone as a result.  Also added a stop
method to the listen socket while I was at it.  The login/logout process
got rearranged as a result of the access pool move.

I was able to get rid of a good number of global variables from the zone
interface file, in anticipation of getting rid of it entirely.  The active
users set wasn't used, and was mostly duplicated data, so it has been
removed.  We can add it, or something like it, back later if need be.

While working on the login/logout changes, the handling of the username
and password data that gets sent by the user was a bit questionable, as
far as possible overflows.  They're now std::strings, limited by the length
of the fields.