Tiny updates
We're now linking properly and running!  Luckily Mathematica links against
libstdc++, so we were able to find a fix for our program not linking to the
right library.  Another library item within the zone interface, the action
lib suffix was incorrect - that should probably be a local define of some
sort.

We also found that we were never starting any of the socket threads once
they were created, which caused some unhandled exceptions on termination,
both of which have been fixed.