Subserver is now its own binary
There were two ways to go with the stream subserver.  One was moving all
the code into the stream socket class; I did some work along those lines,
and it just didn't seem like the right thing to do.  The other was to
completely separate the subserver stuff out into a different binary, and
exec it once we spawned the new process.  The code was mostly segregated
already, so there was little work to do in making the split.