Moved the zone_interface into server.cc/server.h
The zone_interface files were a remnant of when we were trying to keep
the main server as C and have all the C++ contained within the classes
shared library, but since we're no longer doing that, moving the last
bits into the main server.cc was just the right way to go.  There are a
few points at which either the database object or the zone object are
directly needed, so those includes changed to server.h.