Eliminated logging in modules
There was some questionable linking behaviour in some of the modules
relating to the std::clog/syslog stuff we've been doing, so instead
of taking care of logging before we threw something, we'll now just
create the message and throw a standard exception.  All database and
console modules should be logging-free, and should only throw things.