Templatized the geometry and texture caches
The object cache is now a template.  It takes two or three arguments:
the object type (an arbitrary structure, or container), a parser type,
and an optional cleanup routine struct.  The default cleanup is a no-op,
since all object types will not need cleanup.

The geometry cache has a cleanup routine which goes through each element
in its vector (the 'geometry' type is a vector of OpenGL display list
IDs), and frees up each of the display lists.

Also created a much stripped-down r9client.h, which removes most of the
contents of client.h.