Texture and geometry cache parser cleanups
It seemed like it might have been a good idea to have pre-transcoded
strings hanging out, ready to compare, but it turned out that they didn't
really work as expected.  So they're gone; we just transcode the output
from the parser driver, and do it the uninteresting way.  I really wish
the SAX parser would just use std::strings, but alas.

Some testing, and the texture cache looks good.  The geometry cache is
probably fine, but since there's a lot of GL activity during the parser,
we haven't gotten a good test yet.

Moved some of the "things to do" from the source files into TODO.