Repaired sector allocation in the zone
The init and destructor methods in the zone now properly allocate and
deallocate their sectors without seg-faulting.

The motion and octree classes were also having some trouble finding
symbols in the executable, which turned out to be simply the wrong
includes being used.

The executable starts and attempts to run, but terminates due to
inability to connect to the database server on localhost.  That'll be
the next area of debugging.