Added movement handling into the zone
Since we've now got somewhat-functional octrees, we've updated the zone
to take advantage of the new functionality.  The motion worker will now
properly remove and insert things as necessary, and will even create
new sectors as necessary.

Also changed the top-level Makefile target to do rebuilds on each
submodule in order to build the top-level target.  So if we update
something in classes/ for example, it will be rebuilt, and the top-
level target will also be relinked against it, instead of having to
change into the classes/ directory and remake by hand.