Update server config to config_data
The autoconf-created config.h and our server-config config.h were
conflicting, so we've renamed the server config one to config_data
(which, incidentally, is the actual name of the class).

Also, some updates to the protocol module to eliminate conditional
compiles, and also to make the server_notice packet make more sense.