Lots of somewhat autoconf-related updates
We started working on the various pieces of the client, and making sure
that everything configured and built as expected, but it turned out that
there was some stuff that didn't link correctly, due to functions being
commented or unavailable.

One major item was the refactor of the ObjectCache into the
BasicCache and the derived ParsedCache.  There will probably be some
more work needed by it, but the related files build correctly.