First unit test
We need to add unit tests for as many of the components as we can, so
we're starting with the server Library object.  Fairly simple to mock
and test.  This also includes Google Test and Mock frameworks, and some
automake/autoconf changes to make them go.  We also grabbed a TAP
listener for GTest from https://github.com/kinow/gtest-tap-listener.
First commit for issue #3.