Add test for client cache object
Per issue #3, we need unit tests for as much of the codebase as we can,
so we've moved into the client.  The cache covers most of the stuff
for the moment - the object, geometry, and texture caches all use
this basic object.

Until we merge this branch back into master, we don't have the
necessary calls to be able to add objects into our cache.  We can
deal with that when we get there.