Allow client cache index operator to operate reasonably
While working on unit tests, we discovered that there was not actually
a way to insert items into the BasicCache object.  The index operator
would toss an exception when it didn't find an item, which violates
the semantics of the underlying object, and really objects in general.
So instead, we'll now follow what the unordered_map does, and return
a zero-initialized object reference, which is keyed by our new object
ID.