Get perl language lib to compile once again
Apparently some things have changed, and the perl library wouldn't
compile any more.  I relaxed some of the warnings, and made another
couple of tweaks, and it's in good shape now.  Now we can make some
tests.

Also, set all of the non-library libraries to compile as modules.
That will give us a consistent filename (whatever.so) that we can use
for testing, or loading, or whatever.