A couple fixes to build under Linux
A couple items within the test suite were having some trouble
building or running under Linux, so some quick fixes to resolve
them.

The Library test used a blank name, which fails under OSX, but
is completely legal under Linux.  We're now sending in a bogus
library name, instead of nothing.

The TAP harness needed the algorithm include added.  This might
end up as a bit of back-and-forth, with some configure items,
or at the very least some #ifdef in the source file.