Add unit tests for embedded languages
Issue #22:  We need unit tests for as much of the codebase as
possible, and the embedded languages are a relatively isolated
part, so they can be easily added.  Both unit tests create and
destroy an interpreter, and also create a function to raise a
number to a power.

The perl test is uncovering some problems with the perl
interpreter.  My motivation to fix it is fairly low at the moment,
so I'm putting in the failing test, with a warning in the configure
script that it may be crashy.  The test script that we're using
works fine on the command line, so it's something with the way
we're creating the interpreter.