Final working version of pass_fd
Issue #24 was showing some segfaults under Linux, so we had to do
some tweakery.  Found a good working version at Stack Overflow,
so we're working on both OSX and Linux now.

Also updated the pass_fd in the stream socket class, since it
used the exact same thing that was in the subserver test.