Client configdata test is complete
There were some bad assumptions being made in the configdata, so
those needed to be fixed.  Now the class is cleaner as a result,
and we have a unit test.  Not as unit-y as I'd hoped, as there
should be more mocking, but gmock was being all weird.  I think
we can call issue #20 done.