Update client code for localization
For issue #29, we need the client to be able to localize its output.
There are two parts to this.  The first part is the Xt portion, which
is handled by adding an XtSetLanguageProc call to the top of main().
App-default files will be added in forthcoming commits, as I have
apparently deleted the one I had.

The second part involves the messages we show in the message pane,
which we'll use GNU gettext to handle.  Just two messages for now,
but more will be forthcoming.

We'll add the automake parts of these two pieces in the commits
which first add the relevant files.