Add unit test for the thread_pool
Issue #12 wants us to add a test of the thread pool.  We also
made a couple minor tweaks to the thread pool itself.

We found that pthreads lacks the ability to kill a single thread
by id, or at least not without a lot of scaffolding in place, so
the pools can only grow now.