Add unit test for the Sockaddr objects
For issue #5, we need to do testing for all of the server-side
socket objects.  The Sockaddr family of objects seemed to be the
most straightforward, so that's where we started.

Also a few fixes to the objects themselves.  The less_sockaddr
struct got turned into a std::function, since std::binary_function
is deprecated in C++11.