Reworked basesock port number argument
While testing the basesock, we noticed that we were getting some
redundant information in the arguments - the port number is
contained within the struct addrinfo, so we removed the separate
port argument.

We also put our lovely sockaddr objects to better use, which allowed
us to remove some verbose conditionals.  We added a couple access
methods to the sockaddrs, to simplify things for callers.