Add test for server basesock
This is a unit test for the basesock server socket base class.  Re:
issue #5.

There is also a base_user in the same file, but it hooks together
a lot of parts of the server, so mocking everything is a little
ridiculous.  We'll just hold off on testing that one for now.