Changes for basesock test on Linux
The basesock test was running fine on OSX, but was segfaulting on
Linux.  One of the problems was that we were entering into stop()
without a valid running thread; we've now got a private bool to
flag us that we've run start().

Second was that the bad-constructor test was also segfaulting.
That test doesn't really test anything useful, so I just removed
it.  Passing with flying colors now.

Re: issue #5.