Added Sockaddr_un, refactored basesock to understand
While looking at the consoles for unit testing, we decided that since
the console itself was largely a copypasta of the basesock, that we
should just *use* the basesock.  We do need to support unix-domain
sockets for the console, so we needed to get basesock to understand
them.

As a side-effect, we might be able to support unix-domain listening
sockets in the main server.  They probably wouldn't be supremely
useful, but would be amusing.