Moved the base_user into listen_socket files
Since the base_user is more useful for the listen_socket, it makes
much more sense for them to be together.  Also, since we're going
to be using the basesock for the consoles as well, and base_user
doesn't make any sense for them, we shouldn't pollute the memory
space of the console module.