The big console refactor
In order to get the consoles into a testable state, they needed some
drastic refactoring.

First, we needed a portable way to get a hostname from an IP, so the
Sockaddr family of objects now has the hostname() method.  The
Sockadd_un just returns a static "localhost".  For some reason,
the less_sockaddr functor wasn't playing nice with the rest of the
new includes, so that got moved into the dgram server socket,
where it's actually used.

Next, since we are fully generalized with the address handling, we
were able to copy the interesting parts of the sub-console types
(InetConsole and UnixConsole) into the Console itself, and get rid
of the subtypes.

Finally, the Console itself is derived directly from the basesock.
We add our couple new methods, and that's really about it.

While we were at it, the basesock had some residual base_user stuff
which we moved into the listensock.h header.