Small change to allow the Sockaddr_in test to succeed
The memcmp we had in the less-than method was failing to test
correctly (it may be the case that memory was not being initialized
as we might have expected).  Since the values were correct, we
changed the method to compare the values, and the unit test works
as expected once again.