Small test fixes
Apparently a number of our tests have broken for weird reasons, or
they weren't working correctly on the Linux host in the first place.

The less method of the Sockaddr_in wasn't working correctly because
of the network/host byte ordering.  Added some ntohl and ntohs calls
and the test is working as expected.

The libwrap was being applied to everything, which wasn't what we
want, so we put some guards and AC_SUBST calls into the configure.ac,
so that WRAP_LDLIBS is now a thing.