Repairing some autotools weirdness
Apparently in the course of working on the unit tests, some of the
previously-working stuff had stopped working.  The libtest.so test
library was no longer being built as a shared library, since
automake and libtool assume (correctly) that nothing in the test
suite will ever be installed, and (incorrectly) that any test lib
should never need to be a shared lib.  So we build a regular lib,
and set a local install-bin rule to not do anything.

The t_sockaddr test has a mocked getnameinfo, and the prototype is
different between Linux and OSX, so sometimes the test wasn't even
compiling.  We've now got a test within the configure which checks
out the type of the flags argument, and allows the test program to
succeed.