Small cleanups on console unit test
The gtest ASSERTs within the mocked getnameinfo function were not
acting very well, so instead we now just check things the old
fashioned way.

We do need to add a mechanism to connect to the socket and send and
receive data, but I think I'll leave that to when we test the main
server sockets.  We should be able to reuse the same program here.

This should conclude issue #16.