Add entity resolver for DTD files
The geometry and texture objects depend on the ability to parse
XML files of the expected types.  To do that, of course, we need
doctype definitions, and we need to be able to find them in the
filesystem.  Default behaviour is to look in the current directory,
which is unrealistic given that we may have many directories which
contain the relevant data files.  So we've made an entity resolver
which will look for the relevant .dtd files within the datadir
that we create on install.  Re: issue #25.