Unit test for DTD resolver
There does not currently exist a unit test for the texture parser
(or any parser for that matter), so we're starting one in order to
test our new DTD resolver.  We didn't get everything right with
regard to the implementation, so fixes to get it to build are here.
Unfortunately the test segfaults at the moment, so more work is
required.