Small build changes
Our Linux dev host was recently updated to CentOS 7, and some of
the files needed some small tweaks.  Autoconf was updated, so there
were some small changes to config.h.in associated with that.

In order to make the test tree not install anything, we basically
have to error out.  So the install-exec-local rule now does that.

We also ran into some missing or incorrect includes within the
stream socket.  We added some configure checks for the needed
header files, and conditionals around the actual includes.

Console test was running into a missing define, which was resolved
by just including the tcpd.h header.