Update queue now sends updates
In order to try getting pixels on the screen, we're now sending
position updates out of the server.  We're using the send_update
method of the control object, instead of making one packet and
sending that out to everyone.  We may yet want to go back to that
mechanism, once we get enough objects moving around.

Re: issue #8.