Factored the update pool into its own class
It seems that the simplest way to do the various thread pools
within the zone is to make them their own classes, so we can
instantiate them directly, and mock the surroundings.  UpdatePool
is now separate from the Zone, and simplifies the actions within
the Zone as well.

The other two thread pools within the zone will be refactored in
their own branches.