Factored the action pool into its own class
In testing the various thread pool workers within the zone, we
determined that they needed to have less stuff around them, to be
able to easily test them.  So we're factoring them out as we start
writing tests for each.

Unfortunately the action pool needs access to some other stuff
within the zone, so we can't just keep it self-contained; the
startup arg needs to be the zone pointer, rather than the pool
pointer.

Re: issue #13.