Factored sending out of Control
The Control object has no business sending packets, so all of the
sending stuff is gone, and moved into the listening socket objects.
The changes cascaded into a number of different files, so there
are a lot of small changes in a bunch of different files.

First start on the server/subserver protocol, and it's getting
there, but we'll need to work out what happens when the server
passes in a file descriptor - the subserver uses a different file
descriptor number than the server does, so we'll need to send that
back up somehow.