Removed TAP test listener
Since the TAP test listener didn't understand gmock, and seemed to
give *less* data, we've removed it.  Originally I thought it would
give assertion-by-assertion output as with perl's Test::More, but
it didn't.

Perhaps we can try it another time, with some wrappers around the
various gtest assertions, to give the sort of output we actually
do want.