Repaired a small error in listen_socket
In writing the test for the update pool, we saw in our log that
the listen_socket tries to kill its reaper thread, even if it
hasn't been started up.  We've added a boolean guard around the
cancel/join, and set the boolean true in the dgram and stream
sockets when we crank up the reaper thread.