Unit test for the update pool
Now that we've got a test suite, every new addition needs a unit
test.  The update pool is split out from the zone, in order to make
testing a little easier.  I'm not sure if the goal was actually
hit, but this is a functional test, which succeeds.

This should close out issue #15.