Cleaned up old object cache
We had a by-hand object cache which lived within the GL viewport
widget, and was no longer necessary, so it's now gone.

We did a surprising amount of work to replicate old behaviour, so
that the GL viewport should just actually work now.  We'll do some
integration testing to see if anything actually gets drawn.  This
is in service of issue #52.