Add basic geometry and texture datastore files
We need to supply a very basic geometry (a sphere), and a very basic
texture (shiny medium grey) with the client for a fallback for each
of the caches.  When we can't find an ID in the caches, we'll return
these basic elements until we can load up the actual requested ones.