Small updates for MySQL database module
We hadn't been building the MySQL module, and it was having some
significant problems due to file naming - the mysql.h that MySQL
provides was being hidden by the mysql.h in our files.  Renaming
them to mysql_db.h and mysql_db.cc solved that.  There was also
a left over change from the GameObject refactor that didn't make
it into this module that's now fixed.