Cleanup of the dgram socket receive worker
In the course of getting everything working for issue #52 testing,
we found that the dgram socket's receive worker was getting a lot
of weird half-data.  We're now just dropping that stuff on the
ground, and only proceeding when we have legitimate packets from
legitimate hosts.

We still need to test this object to make sure it is working as we
think it is.