Clean up client comms and get things hooked up
For issue #52, we need the client comms to at least work.  They
previously wouldn't even be created.  They're created, started,
and we send logins and logouts at the appropriate times.  Nothing
crashes or errors out.  Theoretically, they're working, but the
server is having some difficulties at the moment, so we're not
sure.