Fixed install rules for geometry and texture files
We were passing a specific flag to the stat program in the install
rules for both geometry and texture files, but neither was correct.
It turns out that the flags don't matter, since we're eating all
the output anyway.  Removed the flags, and the installs do exactly
what we wanted in the first place:  only makes directories which
don't exist, and only installs files which need to be installed.