Add CLO to not detach from controlling tty
Issue #54 requests a command-line option to cause the server not
to detach from its controlling tty, for purposes of debugging -
most debuggers can't follow the child process of a fork.  We've
taken the extra step of not opening the syslog and just printing
the log entries to stderr (the usual destination for std::clog)
when we're in no-detach mode.