Rework of client Comm object
We're working on getting the client and server talking to each
other, and it seemed like the client comm wasn't doing very well.
Instead of doing copying, we'll just pass around pointers to
the packet structures.

We also missed a couple of small things in the test makefile for
issue #54.