Update actions to return a status
In issue #48, we needed to be able to return an acknowledgement to
a client for any given action.  We simply return an int, which is
then evaluated by the zone, and if we get a value that warrants a
return, we send out the ack packet.

Values of:
* -1 : an error has occurred, or some object doesn't exist
* 0  : success
* >0 : we need to return something to the client

We don't currently do anything for the error case, but could in
the future.

This required a fairly invasive refactor of the packet_list struct.
We ended up casting a uint64_t to a Control object almost all the
time, so we just make it a pointer to a Control.  We also needed
the ability to know where a request came from, so we added a
pointer to the relevant listen_socket object.