Updated login-related queries in MySQL module
Now that we're actually accepting logins, and the server seems to
be running pretty well, we messed with the database structure and
actions a bit.  The open_new_login and close_open_login methods
were unimplemented, so we finished them up.

We needed the server ID value in a couple of the login-related
methods, so we've added a member to keep track of it, and within
the MySQL object, we're retrieving it in the connect routine if
it's not already set.

The default "no access" value for checking server authorization is
now 1, rather than 0, so that will be our default return value.

Also some small tweaks to the MySQL schema based on the results of
some testing.