Get client ready to draw things on the screen
Issue #52 has us getting the client ready to draw things on the
screen.  We had previously removed the draw_geometry and
draw_texture routines, in order to get things compiling, but we
need them back.  It made more sense to put in a "client core",
which contained the caches, so we wouldn't have crazy includes and
extern declarations all over the place.

We may yet move the comm object into the client core.