Add GLUT initialization
Once we got the client loading up its default objects, which
includes a sphere that gets drawn by GLUT, we were getting
complaints that GLUT wasn't being initialized.  The view now takes
care of that initialization.  We've also moved the client core
initialization into the view's init callback, so we're guaranteed
that everything is ready to go before we start trying to use it.