Hook a newly-logged-in user up to their game object
When we login, we're now logging in with a character name as well,
and we'll get hooked up to our character object immediately on
login.  That's handled by the zone.  This is to get things ready
for issue #57.

While we were doing that, we decided to typedef all the iterators
to the various containers throughout the server classes, since the
names can get pretty long.  Now things are much easier to read, not
to mention more descriptive - instead of having some random
iterator type, we now see what class it's from, and have a better
idea what it's doing.