Send position updates for everything on login
For issue #57, we need to send position updates for everything
nearby when somebody logs in, so they can actually have something
to look at.  This isn't ideal, as the updates will get sent to
everyone, but only the stuff within 1000m of the recently logged-
in user.

The game object will now compute its distance from a given point,
which should be useful in more than just this situation.