Rename server_objects database table
We need a way to save data for non-character objects in the
database.  Instead of adding a second table, we can just add the
functionality to the existing server_characters table.  We're
renaming it to server_objects, and tweaking the importance of the
objectid field slightly.  We're also adding a geometryid field,
for the time when that becomes important.