Stop saving password in client settings
Saving a user's password in plaintext is admittedly a very bad way
to go, so we're no longer saving it at all.  Instead, we'll pop up
a dialog to get final login params, including the password.  The
text field that takes the password echoes *s instead of the actual
characters.

Re: issue #59