Return correct access type on login
We were not returning a correct access type to the client upon
login.  Issue #61 needs us to do so.  We added a second database
call to fetch the authorization via character name, in addition to
the one that takes a character id value.

The listen socket will now create (or unmark for reaping) a control
object for any valid username/password pair, even if the access
level is ACCESS_NONE.  We need the control object to be able to
send something back to the user in question, even if they're
disallowed.  If access level is ACCESS_NONE, each of the derived
socket classes will mark the control for reaping in the do_login
method.  Otherwise, the user will be logged in, and things will
continue forward as before.