Receive tests for client comm object
The receive half of the unit test for the client comm object is
done.  Sending tests to come next.  Re: issue #17.

The mocking was a little tricky, and there may be some race
conditions on heavily loaded systems.  The receive thread starts
up in the constructor, so we don't have any reasonable time when
we can set the EXPECT_CALL items for gmock.  In order to give
ourselves a window, we have our mock recvfrom function sleeping
for one second, so that we have time to set the expectations.
It is definitely not ideal, but it does work for the time being.

In the future we may consider having start and stop methods, so
we can control the thread pools better, and don't have to resort
to this ridiculousness.