Add start/stop methods to client comm
During testing of the client comm object, we found that when we had
to mock out the receive methods, there was no seam we could use to
set gmock's expectation following object creation.  Since most of
the rest of our thread-pool-like objects have start and stop
methods, we added them here to allow ourselves time to set the
expectations we need.