Delete packets we can't hton
In issue #63, we find that we will leak memory for each packet that
we can't properly hton.  The simplest change is to delete the
packet at that point, and not worry about returning anything.  The
comm object is the only caller of the send() method, and if we're
generating bad packets there, they'll *all* be bad, and we'll need
to make a code change to fix the problem.

For the most part, the only reason an hton call will fail is when
we pass a bad size in for the packet type we're operating on.