Generate some shaders
We're going to get anything on the screen by any means necessary.
For the time being, we'll just have points, which we'll draw as
boxes on the screen.  This commit gets some of the basic functions
in place for making shaders, creating shader programs, and setting
up a projection matrix for our window.  It almost certainly does
not work, but it does compile and run.

Re: issue #52.