Add orientation handling to the geometry handling
As we move on with issue #52, the geometry shader didn't have any
facility to handle orientation of the boxes it would draw.  We do
send those values from the server, so we'll go ahead and use them.

We completely removed any GL handling in the geometry parser.  It
simply doesn't need to be here.  We could probably remove the
geometry and texture parsers in their entirety, since we're just
using points which get transformed into boxes, and a single
texture.

Also, the client comm test didn't get updated following a previous
change, and was segfaulting with a double-free error.