Cleaned up testing
The t_texture_parser test is no longer meaningful, because the
texture object no longer exists in this branch.  So it's gone.

We've just reverted the modification to gtest/gmock to turn off
the fused-src tests, because we figured out how to turn off
recursion into the gtest/gmock tree for 'make check'.  Turns out
you can have per-target variable changes, so we've turned off
SUBDIRS for the check target in test/Makefile.am.