Implemented full drawing process
We've now got a VBO for each object, which contains an XYZ point,
and an RGB color, and we should be drawing each point, with proper
orientation.

We made the assumption that since all the VBOs are arranged the
same, we should be able to set the attribute locations once, at
program creation time, and we shouldn't need to make those calls
again.  That may not be the case, of course, so we shall see.