OpenGL updates
In our efforts for issue #52, and a bit of printf-debugging, we've
found some errors and done some cleanups.

The semicolons after all the main shader programs were syntax
errors, so the shaders *should* compile now.  The Xquartz GLX
extension only offers OpenGL 2.1, so the programs actually fail
to compile with an unsupported version.

We made the shader and program creation routines throw exceptions
on failure, so we're not left in some funky state.  Another utility
"turn a GLenum into a readable string" function, to improve those
error messages.

Additional nice output.  It's nice!